requirements.txt
# Core
#torch>=2.0.0
#torchvision>=0.15.0

# Communication
grpcio>=1.54.0
grpcio-tools>=1.54.0

# Web Dashboard (optional but nice)
flask>=2.3.0
flask-socketio>=5.3.0

# Utils
pyyaml>=6.0
python-dotenv>=1.0.0
psutil>=5.9.0  # For system monitoring
numpy>=1.24.0
requests>=2.31.0

# For nice CLI
colorama>=0.4.6
tqdm>=4.65.0